--- apps/users/models/__init__.py
@@ -1,1 +1,0 @@
-from .custom_user import CustomUser
